Increase max TravelTime passes, but stop sooner.
The Google Distance Matrix API tends to return pretty stable results for
far-off distant events, or for off-peak hours.

As such, we can reduce the number of API calls by detecting this early
and skipping any remaining passes.

With this in mind, we can also increase the accuracy for nearby or
time-sensitive events (e.g. rush hour) by increasing the max pass count.