Retrieve travel times directly from events.
This replaces our Distance Matrix API calls with CalDAV API calls
instead, using the X-APPLE-TRAVEL-DURATION tag.

I've got tentative plans to bring back the Distance Matrix API in a
separate project, which would calculate (pessimistic, buffered) travel
times for all events, refresh them as they approach, and possibly change
the origin based on the user's location if the event is getting close
and they're not at the standard origin — whereupon I could use the
Porsche "position" call as (one possible source of) user location data.

If/when I make that project, this project will be able to read those
times and react accordingly.  But I don't think we need all that extra
logic in this project (even if there would be some duplication of code
between the two), and a "travel time auto-calculator" project shouldn't
be tightly tied to the Porsche API like this project is.